Improve description of the secion, move to dedicated page
